Adding vnc, Fixing snap/take.c

+++ sys/src/cmd/vnc/vncv.h
@@ -9,7 +9,7 @@
 
 /* color.c */
 extern	void		choosecolor(Vnc*);
-extern	void		(*cvtpixels)(uchar*, uchar*, int);
+extern	void		(*cvtpixels)(uint8_t*, uint8_t*, int);
 extern  void            settranslation(Vnc*);
 
 /* draw.c */
@@ -17,7 +17,7 @@ extern	void		sendencodings(Vnc*);
 extern	void		requestupdate(Vnc*, int);
 extern	void		readfromserver(Vnc*);
 
-extern	uchar	zero[];
+extern	uint8_t	zero[];
 
 /* vncv.c */
 extern	char		*encodings;
@@ -31,5 +31,5 @@ extern	void		initmouse(void);
 extern	void		mousewarp(Point);
 extern	void		readmouse(Vnc*);
 extern  void            senddim(Vnc*);
-extern  void            writesnarf(Vnc*, long);
+extern  void            writesnarf(Vnc*, int32_t);
 extern  void            checksnarf(Vnc*);
